+++ packages/caption/src/styles/index.android.js
@@ -10,11 +10,12 @@ const styles = StyleSheet.create({
   },
   credits: {
     ...sharedStyles.credits,
-    lineHeight: 20
+    lineHeight: spacing(4),
+    marginTop: 0
   },
   text: {
     ...sharedStyles.text,
-    lineHeight: 20
+    lineHeight: spacing(4)
   }
 });