+++ packages/slice-layout/src/templates/secondarytwonopicandtwo/styles.js
@@ -1,13 +1,22 @@
 import { spacing } from "@times-components/styleguide";
 
 const mediumBreakpointStyles = {
-  item: {
-    width: "50%"
+  container: {
+    flex: 1,
+    marginHorizontal: spacing(4)
   },
   itemContainer: {
     flex: 1,
-    flexDirection: "row",
-    marginHorizontal: spacing(2)
+    flexDirection: "row"
+  },
+  item: {
+    flex: 1
+  },
+  secondaryColSeparator: {
+    marginTop: spacing(3)
+  },
+  supportColSeparator: {
+    marginBottom: spacing(3)
   }
 };