+++ android-app/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.26](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.25...@times-components/android-app@0.15.26) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.25](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.24...@times-components/android-app@0.15.25) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.24](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.23...@times-components/android-app@0.15.24) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.23](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.22...@times-components/android-app@0.15.23) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.22](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.21...@times-components/android-app@0.15.22) (2019-08-02)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.21](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.20...@times-components/android-app@0.15.21) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.20](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.19...@times-components/android-app@0.15.20) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
+## [0.15.19](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.18...@times-components/android-app@0.15.19) (2019-07-30)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
 ## [0.15.18](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.15.17...@times-components/android-app@0.15.18) (2019-07-30)
 
 **Note:** Version bump only for package @times-components/android-app