+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.17](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.16...@times-components/article-image@7.3.17) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
+## [7.3.16](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.15...@times-components/article-image@7.3.16) (2019-07-31)
+
+
+### Bug Fixes
+
+* paragraph with dropcap margin bottom fixed ([#2122](https://github.com/newsuk/times-components/issues/2122)) ([f76c597](https://github.com/newsuk/times-components/commit/f76c597))
+
+
+
+
+
+## [7.3.15](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.14...@times-components/article-image@7.3.15) (2019-07-30)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
 ## [7.3.14](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.13...@times-components/article-image@7.3.14) (2019-07-29)
 
 **Note:** Version bump only for package @times-components/article-image