+++ packages/article-in-depth/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.9.12](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.11...@times-components/article-in-depth@3.9.12) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.9.11](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.10...@times-components/article-in-depth@3.9.11) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.9.10](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.9...@times-components/article-in-depth@3.9.10) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.9.9](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.8...@times-components/article-in-depth@3.9.9) (2019-08-02)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.9.8](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.7...@times-components/article-in-depth@3.9.8) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.9.7](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.6...@times-components/article-in-depth@3.9.7) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
+## [3.9.6](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.5...@times-components/article-in-depth@3.9.6) (2019-07-30)
+
+**Note:** Version bump only for package @times-components/article-in-depth
+
+
+
+
+
 ## [3.9.5](https://github.com/newsuk/times-components/compare/@times-components/article-in-depth@3.9.4...@times-components/article-in-depth@3.9.5) (2019-07-30)
 
 **Note:** Version bump only for package @times-components/article-in-depth