+++ packages/depend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.21](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.20...@times-components/depend@2.1.21) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/depend
+
+
+
+
+
 ## [2.1.20](https://github.com/newsuk/times-components/compare/@times-components/depend@2.1.19...@times-components/depend@2.1.20) (2019-05-28)
 
 **Note:** Version bump only for package @times-components/depend