+++ packages/error-view/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.83](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.2.82...@times-components/error-view@2.2.83) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/error-view
+
+
+
+
+
 ## [2.2.82](https://github.com/newsuk/times-components/compare/@times-components/error-view@2.2.81...@times-components/error-view@2.2.82) (2019-07-29)
 
 **Note:** Version bump only for package @times-components/error-view