+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.5](https://github.com/newsuk/times-components/compare/@times-components/icons@2.12.4...@times-components/icons@2.12.5) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/icons
+
+
+
+
+
 ## [2.12.4](https://github.com/newsuk/times-components/compare/@times-components/icons@2.12.3...@times-components/icons@2.12.4) (2019-07-29)
 
 **Note:** Version bump only for package @times-components/icons