+++ packages/link/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.33](https://github.com/newsuk/times-components/compare/@times-components/link@3.4.32...@times-components/link@3.4.33) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/link
+
+
+
+
+
 ## [3.4.32](https://github.com/newsuk/times-components/compare/@times-components/link@3.4.31...@times-components/link@3.4.32) (2019-07-29)
 
 **Note:** Version bump only for package @times-components/link