+++ packages/markup-forest/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.17](https://github.com/newsuk/times-components/compare/@times-components/markup-forest@1.7.16...@times-components/markup-forest@1.7.17) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/markup-forest
+
+
+
+
+
 ## [1.7.16](https://github.com/newsuk/times-components/compare/@times-components/markup-forest@1.7.15...@times-components/markup-forest@1.7.16) (2019-07-12)
 
 **Note:** Version bump only for package @times-components/markup-forest