+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.93](https://github.com/newsuk/times-components/compare/@times-components/markup@3.3.92...@times-components/markup@3.3.93) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
+## [3.3.92](https://github.com/newsuk/times-components/compare/@times-components/markup@3.3.91...@times-components/markup@3.3.92) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
+## [3.3.91](https://github.com/newsuk/times-components/compare/@times-components/markup@3.3.90...@times-components/markup@3.3.91) (2019-07-30)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
 ## [3.3.90](https://github.com/newsuk/times-components/compare/@times-components/markup@3.3.89...@times-components/markup@3.3.90) (2019-07-29)
 
 **Note:** Version bump only for package @times-components/markup