+++ packages/pages/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.142](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.141...@times-components/pages@2.0.142) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.141](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.140...@times-components/pages@2.0.141) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.140](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.139...@times-components/pages@2.0.140) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.139](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.138...@times-components/pages@2.0.139) (2019-08-05)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.138](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.137...@times-components/pages@2.0.138) (2019-08-02)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.137](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.136...@times-components/pages@2.0.137) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.136](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.135...@times-components/pages@2.0.136) (2019-07-31)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
+## [2.0.135](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.134...@times-components/pages@2.0.135) (2019-07-30)
+
+**Note:** Version bump only for package @times-components/pages
+
+
+
+
+
 ## [2.0.134](https://github.com/newsuk/times-components/compare/@times-components/pages@2.0.133...@times-components/pages@2.0.134) (2019-07-30)
 
 **Note:** Version bump only for package @times-components/pages